# src/app/api/openapi.yaml
openapi: 3.0.0
info:
  title: ZenZen CRM API
  version: '1.0.0'
  description: |
    API for the ZenZen CRM system, providing ticket management, messaging, and team collaboration features.
  contact:
    name: ZenZen Support
    url: https://zenzen.support
servers:
  - url: https://api.zenzen.dev
    description: Development server
  - url: https://api.zenzen.com
    description: Production server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    WorkerChatMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        worker_chat_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        content:
          type: string
        created_at:
          type: string
          format: date-time
...
paths:
...
  /worker-chat/{chat_id}/messages:
    parameters:
      - name: chat_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get messages in a worker chat room
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of chat messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkerChatMessage'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      pages:
                        type: integer
                      current_page:
                        type: integer
                      per_page:
                        type: integer

    post:
      summary: Send a message in worker chat
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
      responses:
        '201':
          description: Message sent successfully
...
  /webhooks:
    post:
      summary: Register a new webhook
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
                - events
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                    enum: [ticket.created, ticket.updated, message.created, feedback.submitted]
                secret:
                  type: string
                  description: Secret for webhook signature verification
      responses:
        '201':
          description: Webhook registered successfully
...
security:
  - ApiKeyAuth: []
  - BearerAuth: [] 